DDP-5667: Move question prompt component to a separate directory
