+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/answers/question-prompt/questionPrompt.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ActivityQuestionBlock } from '../../../models/activity/activityQuestionBlock';
+import { ActivityQuestionBlock } from '../../../../models/activity/activityQuestionBlock';
 
 @Component({
     selector: 'ddp-question-prompt',
